+++ client/src/main/js/controller/AgendaController.js
@@ -39,8 +39,7 @@
             this.clickDia = function (data) {
                 const calendario = uiCalendarConfig.calendars.calendario;
 
-                if (calendario.fullCalendar('getView').type === 'agendaDay' ||
-                    calendario.fullCalendar('getView').type === 'agendaWeek') {
+                if (isVisualizacaoDiaOuSemana(calendario.fullCalendar('getView'))) {
                     if(data.isAfter(moment())){
                         self.criarReserva(data.toDate());
                     }
@@ -50,15 +49,27 @@
             };
 
             /**
+             * Verifica se o calendário está na visualização da semana ou do dia.
+             *
+             * @param view Tela em que o calendário está renderizado.
+             * @return {boolean} {@code true} se o calendário está na visualização
+             * da semana ou do dia, {@code false} caso contrário.
+             */
+            function isVisualizacaoDiaOuSemana(view) {
+                return view.type === 'agendaDay' || view.type === 'agendaWeek';
+            }
+
+            /**
              * Objeto enviado à diretiva do calendário com todas as configurações desejadas.
              * Consultar: https://fullcalendar.io/docs/
              */
             this.calendarioConfig = {
                 calendario: {
-                    height: "100%",
                     editable: false,
                     ignoreTimezone: false,
                     timezone: 'local',
+                    minTime: self.local.getInicioFuncionamento(),
+                    maxTime: self.local.getFimFuncionamento(),
                     lang: 'pt-br',
                     header: {
                         left: 'month agendaWeek agendaDay',
@@ -73,7 +84,7 @@
                     },
                     eventClick: self.clickReserva,
                     dayClick: self.clickDia,
-                    eventAfterAllRender: transformaBotoesCalendario,
+                    eventAfterAllRender: configCalendarioPosRenderizacao,
                     buttonText: {
                         agendaWeek: 'Semana',
                         agendaDay: 'Dia'
@@ -134,6 +145,35 @@
                     });
                 }
                 return [];
+            }
+
+            /**
+             * Executa as configurações que necessitam que o calendário tenha sido
+             * renderizado.
+             *
+             * @param calendarioConfig Objeto que contém informações sobre o calendário
+             * renderizado.
+             */
+            function configCalendarioPosRenderizacao(calendarioConfig) {
+                ajustaAltura(calendarioConfig);
+                transformaBotoesCalendario();
+            }
+
+            /**
+             * Ajusta a altura do calendário, pós-renderização, baseado no tipo da
+             * visualização do mesmo.
+             *
+             * @param calendarioConfig Objeto que contém informações sobre o calendário
+             * renderizado.
+             */
+            function ajustaAltura(calendarioConfig) {
+                const calendario = uiCalendarConfig.calendars.calendario;
+
+                if (calendario) {
+                    const altura = isVisualizacaoDiaOuSemana(calendarioConfig) ? 'auto' : '100%';
+
+                    calendario.fullCalendar('option', 'height', altura);
+                }
             }
 
             /**